now also works with float

+++ src/yars/defines/types.h
@@ -4,7 +4,8 @@
 # define __Y_REAL_MIN -100000.0
 # define __Y_REAL_MAX  100000.0
 
-typedef double yReal;
+// typedef double yReal;
+typedef float yReal;
 
 typedef unsigned int uint;
